backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
opencv-python==4.10.0.84
numpy==1.26.4
scipy==1.14.0
pyaudio==0.2.14
ultralytics==8.2.87
paddlepaddle==2.6.1
paddleocr==2.7.3
pillow==10.4.0
aiofiles==23.2.1
python-multipart==0.0.9
websockets==13.0
httpx==0.27.2
